RU-AI-GETTER/requirements.txt
pdfplumber
beautifulsoup4
# Para integração com Ollama/MCP
requests
pytesseract
Pillow
# Para validação JSON
jsonschema
# Para interface web de revisão
flask
